bot/requirements.txt
slack_bolt
psycopg2-binary
